PetCare/python-backend/app/requirements.txt
langchain-cli
langchain-community
langchain_google_genai
tavily-python
dotenv
fastapi
langchain_community
pypdf
uvicorn